refactor(ui): extract diplomacy row widgets into part to satisfy file-size gate (Refs #3341)
DiplomacyPanel grew to 723 physical lines, breaching the 700-line
repo.game_widgets_file_size gate. Move _DiplomacyRow and _ActionButton
into a new diplomacy_panel_row.dart part fragment; no behavior change.
